requirements.txt
falcon>=3.1.3
supabase>=2.5.1
python-dotenv>=1.0.1
gunicorn>=22.0.0